Adapt configuration script to existing snapshot

+++ scripts/configure.sh
@@ -3,9 +3,9 @@
 
 # clone .emacs configuration from github
 git clone https://github.com/abailly-iohk/dotfiles ~/dotfiles
-ln -s ~/dotfiles/.emacs ~/.emacs
-ln -s ~/dotfiles/.tmux.conf ~/.tmux.conf
-ln -s ~/dotfiles/.gitconfig ~/.gitconfig
+[[ -L ~/.emacs ]] || ln -s ~/dotfiles/.emacs ~/.emacs
+[[ -L ~/.tmux.conf ]] || ln -s ~/dotfiles/.tmux.conf ~/.tmux.conf
+[[ -L ~/.gitconfig ]] || ln -s ~/dotfiles/.gitconfig ~/.gitconfig
 
 emacs --batch -q -l dotfiles/install.el
 
@@ -15,12 +15,16 @@ ssh-keyscan github.com >> ~/.ssh/known_hosts
 # download public keys of interest
 gpg --keyserver keys.openpgp.org --recv 39AF57FB92B465F8AE6FD1BCCB4571C05D7B9E12 B73C82125079C8FC79666FFA59FAA903C906659A
 
-# clone hydra repositories
-git clone git@github.com:input-output-hk/hydra-node ~/hydra-node
-git clone git@github.com:input-output-hk/cardano-ledger-specs ~/cardano-ledger-specs
-git clone git@github.com:input-output-hk/ouroboros-network ~/ouroboros-network
-git clone git@github.com:abailly-iohk/hydra-sim ~/hydra-sim
-git clone git@github.com:abailly-iohk/plutus ~/plutus
+# clone or update hydra repositories
+for repo in "hydra-node cardano-ledger-specs ouroboros-network hydra-sim plutus"; do
+    if [[ ! -d "~/$repo" ]]; then
+        git clone "git@github.com:input-output-hk/$repo"
+    else
+        pushd "~/$repo"
+        git pull
+        popd
+    fi
+done
 
 # configure nix stuff
 source /etc/profile.d/nix.sh
